requirements.txt
django-appconf
django-model-utils
django-sortedm2m
django-taggit
python-dateutil
requests

